requirements.txt
google-genai>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
flask>=2.0.0
flask-cors>=3.0.0